+++ l10n/taot_el.ts
@@ -4,13 +4,13 @@
 <context>
     <name>--------</name>
     <message>
-        <location filename="../src/main.cpp" line="244"/>
+        <location filename="../src/main.cpp" line="254"/>
         <source>AUTHORS</source>
         <extracomment>A list of translation authors</extracomment>
         <translation>Mandratzis-Walz (beonex), Dimitrios Glentadakis (dglent)</translation>
     </message>
     <message>
-        <location filename="../src/main.cpp" line="246"/>
+        <location filename="../src/main.cpp" line="256"/>
         <source>LANGUAGE_NAME</source>
         <extracomment>Native language name (e.g., Deutsch for German)</extracomment>
         <translation>Ελληνικά</translation>
@@ -45,7 +45,7 @@
         <location filename="../qml/bb10/AppCover.qml" line="36"/>
         <location filename="../qml/sailfish/AppCover.qml" line="33"/>
         <source>From: &lt;b&gt;%1&lt;/b&gt;</source>
-        <translation>Απο: &lt;b&gt;%1&lt;/b&gt;</translation>
+        <translation>Από: &lt;b&gt;%1&lt;/b&gt;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/AppCover.qml" line="62"/>
@@ -69,88 +69,89 @@
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="28"/>
         <source>Donation FAQ</source>
-        <translation>Συνήθεις ερωτήσεις Δωρεών</translation>
+        <translation>Συνήθεις ερωτήσεις δωρεών</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="40"/>
         <source>Why do you ask for donations?</source>
-        <translation type="unfinished"></translation>
+        <translation>Γιατί ρωτάτε για δωρεά;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="41"/>
         <source>I&apos;m developing %1 in my spare time and provide it completely for free. Moreover, %1 is Open Source. Receiving donations will encourage me to continue developing and supporting %1. It will also show that my work is appreciated.</source>
-        <translation type="unfinished"></translation>
+        <translation>Αναπτύσσω το %1 στον ελεύθερο χρόνο μου και το παρέχω εντελώς δωρεάν. Επίσης το %1 είναι ανοιχτού κώδικα. Η λήψη μιας δωρεάς θα με ενθάρρυνε στη συνέχιση της ανάπτυξης και της υποστήριξης του %1. Επίσης θα ήταν μια ένδειξη εκτίμησης της εργασίας μου.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="47"/>
         <source>Am I required to donate to use %1?</source>
-        <translation type="unfinished"></translation>
+        <translation>Υποχρεούμαι να κάνω δωρεά ούτως ώστε να χρησιμοποιήσω το %1;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="48"/>
         <source>Absolutely not. %1 is free and Open Source. You&apos;re not required to make any donations to use it. However, this way you can show your support.</source>
-        <translation type="unfinished"></translation>
+        <translation>Απολύτως όχι. Το %1 είναι ελεύθερο λογισμικό ανοιχτού κώδικα. Δεν απαιτείται καμιά δωρεά για την χρήση του. Ωστόσο με αυτόν τον τρόπο μπορείτε να δείξετε την υποστήριξή σας.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="53"/>
         <source>What are those coins?</source>
-        <translation type="unfinished"></translation>
+        <translation>Τι είναι αυτά τα κέρματα;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="54"/>
         <source>I&apos;m using BlackBerry Payment Service to accept donations and BlackBerry takes 30% commission from each payment. For example, if you donate 4.99$ (34 coins), I will get about 3.49$. So each coin roughly corresponds to 0.10$ that I will get.</source>
-        <translation type="unfinished"></translation>
+        <translation>Χρησιμοποιώ την υπηρεσία πληρωμών του BlackBerry για την λήψη των δωρεών και η BlackBerry λαμβάνει ένα ποσοστό των 30% από κάθε πληρωμή. Για παράδειγμα, αν δωρίσετε 4.99$ (4.46€) (34 κέρματα), θα λάβω περίπου 3.49$ (3.12€). Ως εκ τούτου κάθε κέρμα αντιστοιχεί σε 0.10$ (0.08€) που λαμβάνω.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="60"/>
         <source>Do I get something for making a donation?</source>
-        <translation type="unfinished"></translation>
+        <translation>Θα λάβω κάτι αν κάνω μια δωρεά;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="61"/>
         <source>You already have %1. Completely for free :-)
 However, if it&apos;s not enough, the amount of your donation along with &quot;Thank you!&quot; message will be displayed at the top of the About page.</source>
-        <translation type="unfinished"></translation>
+        <translation>Έχετε ήδη το %1. Εντελώς δωρεάν :-)
+Ωστόσο, αν δεν είναι αρκετό, το ποσό της δωρεάς παράλληλα με τις ευχαριστίες θα εμφανίζονται στην κορυφή της σελίδας Περί της εφαρμογής.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="67"/>
         <source>I want to increase my donation. Is it possible?</source>
-        <translation type="unfinished"></translation>
+        <translation>Θέλω να αυξήσω την δωρεά μου, είναι εφικτό;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="68"/>
         <source>Just make a new one. You can make as many donations as you want. They will accumulate and total amount will be displayed on the About page.</source>
-        <translation type="unfinished"></translation>
+        <translation>Απλά πραγματοποιήστε μια νέα. Μπορείτε να κάνετε όσες δωρεές επιθυμείτε. Το ποσό θα συσσωρευτεί και θα εμφανίζεται στην σελίδα Περί της εφαρμογής.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="72"/>
         <source>I changed my mind and want to get my donation back!</source>
-        <translation type="unfinished"></translation>
+        <translation>Μετάνιωσα και θέλω να αναιρέσω την δωρεά μου!</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="73"/>
         <source>Very sorry to hear that. I&apos;m not managing any payments, so you&apos;ll have to contact &lt;a href=&quot;%1&quot;&gt;BlackBerry World support&lt;/a&gt; to request a refund.</source>
-        <translation type="unfinished"></translation>
+        <translation>Λυπάμαι πολύ που το ακούω. Δεν διαχειρίζομαι τις πληρωμές και ως εκ τούτου θα πρέπει να επικοινωνήσετε με την &lt;a href=&quot;%1&quot;&gt;Παγκόσμια υποστήριξη BlackBerry&lt;/a&gt; για να απαιτήσετε την επιστροφή των χρημάτων σας.</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="78"/>
         <source>I don&apos;t want to pay in-app. Are there any other ways to donate?</source>
-        <translation type="unfinished"></translation>
+        <translation>Δεν επιθυμώ την χρήση μιας εφαρμογής πληρωμών. Υπάρχει άλλος τρόπος πληρωμής της δωρεάς;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="79"/>
         <source>Yes, there is &lt;em&gt;Donate&lt;/em&gt; link on the About page. Tap it and you will be taken to a web page where you can donate with PayPal, Flattr, or WebMoney.</source>
-        <translation type="unfinished"></translation>
+        <translation>Ναι, υπάρχει ένας δεσμός &lt;em&gt;Δωρεά&lt;/em&gt; στην σελίδα Περί της εφαρμογής. Ακολουθήστε τον και θα μεταβείτε σε μια ιστοσελίδα όπου μπορείτε να δωρίσετε μέσω του PayPal, του Flattr, ή του WebMoney. </translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="84"/>
         <source>My question isn&apos;t answered. How can I get in touch?</source>
-        <translation type="unfinished"></translation>
+        <translation>Η ερώτησή μου δεν απαντήθηκε. Πώς μπορώ να έρθω σε επικοινωνία;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/DonationFaqPage.qml" line="85"/>
         <source>Write me an e-mail. Just swipe from the top and tap &lt;em&gt;Send feedback&lt;/em&gt; in the menu. An e-mail will open with my address pre-filled.</source>
-        <translation type="unfinished"></translation>
+        <translation>Στείλτε μου μια ηλεκτρονική αλληλογραφία. Απλά σαρώστε από την κορυφή και χτυπήστε στο &lt;em&gt;Αποστολή ανάδρασης&lt;/em&gt; από το μενού. Θα ανοίξει η σύνταξη της αλληλογραφίας με την διεύθυνσή μου προσυμπληρωμένη.</translation>
     </message>
 </context>
 <context>
@@ -181,7 +182,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../src/services/jsontranslationservice.cpp" line="47"/>
         <location filename="../src/services/jsontranslationservice.cpp" line="54"/>
-        <location filename="../src/services/jsontranslationservice.cpp" line="65"/>
+        <location filename="../src/services/jsontranslationservice.cpp" line="80"/>
         <source>Couldn&apos;t parse response from the server because of an error: &quot;%1&quot;</source>
         <translation>Δεν ήταν δυνατή η ανάλυση της απόκρισης από το διακομιστή εξαιτίας ενός λάθους: &quot;%1&quot;</translation>
     </message>
@@ -349,7 +350,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
         <location filename="../qml/sailfish/MainPage.qml" line="51"/>
         <location filename="../qml/symbian/MainPage.qml" line="601"/>
         <source>About</source>
-        <translation type="unfinished"></translation>
+        <translation>Περί</translation>
     </message>
 </context>
 <context>
@@ -386,28 +387,29 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/privacy.js" line="25"/>
         <source>Would you like to help improve %1 by enabling application analytics?</source>
-        <translation type="unfinished"></translation>
+        <translation>Θέλετε να βοηθήσετε στην βελτίωση του %1 με την ενεργοποίηση των αναλύσεων της εφαρμογής;</translation>
     </message>
     <message>
         <location filename="../qml/privacy.js" line="29"/>
         <source>Yes, enable full analytics</source>
-        <translation type="unfinished"></translation>
+        <translation>Ναι, επιθυμώ την ενεργοποίηση των πλήρη αναλύσεων</translation>
     </message>
     <message>
         <location filename="../qml/privacy.js" line="30"/>
         <source>Enable only error reporting</source>
-        <translation type="unfinished"></translation>
+        <translation>Να ενεργοποιηθεί μόνο η αναφορά των σφαλμάτων</translation>
     </message>
     <message>
         <location filename="../qml/privacy.js" line="31"/>
         <source>No, I don&apos;t want to help</source>
-        <translation type="unfinished"></translation>
+        <translation>Αρνούμαι να βοηθήσω</translation>
     </message>
     <message>
         <location filename="../qml/privacy.js" line="34"/>
         <source>&lt;p&gt;%1 contains application analytics functionality that can collect information about actions performed in the application, application errors, and some information about your device. This information is processed by a third party analytics service, &lt;a href=&quot;%3&quot;&gt;Amplitude Analytics&lt;/a&gt;. No personally identifiable information is sent. Collected information will be used solely for the purpose of improving %1, fixing errors, and analysing application usage.&lt;/p&gt;
 &lt;p&gt;Please read &lt;a href=&quot;%2&quot;&gt;%1 Privacy Policy&lt;/a&gt; for full information about what data is sent and when. You can also consult End User Information section of &lt;a href=&quot;%4&quot;&gt;Amplitude Analytics Privacy Policy&lt;/a&gt; to see how they deal with the data that we send to them.&lt;/p&gt;</source>
-        <translation type="unfinished"></translation>
+        <translation>&lt;p&gt;Το %1 συμπεριλαμβάνει την λειτουργία αναλύσεων της εφαρμογής η οποία μπορεί να συλλέγει πληροφορίες σχετικά με τις πραγματοποιηθείσες ενέργειες στην εφαρμογή, τα σφάλματα της εφαρμογής, και μερικές πληροφορίες σχετικά με την συσκευή σας. Αυτές οι πληροφορίες αναλύονται από μια τρίτη υπηρεσία παροχής αναλύσεων, την &lt;a href=&quot;%3&quot;&gt;Amplitude Analytics&lt;/a&gt;. Δεν αποστέλλονται προσωπικές πληροφορίες. Η συλλογή των δεδομένων χρησιμοποιείται μόνο για την βελτίωση του %1, την διόρθωση σφαλμάτων, και την ανάλυση της χρήσης της εφαρμογής.&lt;/p&gt;
+&lt;p&gt;Παρακαλώ διαβάστε την &lt;a href=&quot;%2&quot;&gt;Πολιτική ιδιωτικού απορρήτου του %1&lt;/a&gt; για να ενημερωθείτε πλήρως σχετικά με τον τύπο των δεδομένων που αποστέλλονται και πότε. Μπορείτε επίσης να επισκεφθείτε την ενότητα πληροφοριών του Τελικού Χρήστη στην &lt;a href=&quot;%4&quot;&gt;Πολιτική ιδιωτικού απορρήτου της Amplitude Analytics&lt;/a&gt; για να δείτε πώς διαχειρίζονται τα δεδομένα που λαμβάνουν.&lt;/p&gt;</translation>
     </message>
 </context>
 <context>
@@ -477,7 +479,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/bb10/SettingsPage.qml" line="74"/>
         <source>Hold &lt;em&gt;Shift&lt;/em&gt; while pressing &lt;em&gt;Enter&lt;/em&gt; to start a new line</source>
-        <translation type="unfinished"></translation>
+        <translation>Διατηρείστε πατημένο το &lt;em&gt;Shift&lt;/em&gt; ενώ πιέζετε το &lt;em&gt;Enter&lt;/em&gt; για την δημιουργία μιας νέας γραμμής</translation>
     </message>
     <message>
         <location filename="../qml/bb10/SettingsPage.qml" line="85"/>
@@ -492,7 +494,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
         <location filename="../qml/harmattan/SettingsPage.qml" line="94"/>
         <location filename="../qml/symbian/SettingsPage.qml" line="98"/>
         <source>Automatically start translation after inserting text with &lt;em&gt;Paste&lt;/em&gt; button</source>
-        <translation type="unfinished"></translation>
+        <translation>Αυτόματη εκκίνησης της μετάφρασης κατά την εισαγωγή του κειμένου με το κουμπί &lt;em&gt;Επικόλλησης&lt;/em&gt;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/SettingsPage.qml" line="96"/>
@@ -507,7 +509,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
         <location filename="../qml/sailfish/SettingsPage.qml" line="73"/>
         <location filename="../qml/symbian/SettingsPage.qml" line="87"/>
         <source>Only one line of text is supported in this mode</source>
-        <translation type="unfinished"></translation>
+        <translation>Σε αυτήν την λειτουργία υποστηρίζεται μόνον μια γραμμή κειμένου</translation>
     </message>
     <message>
         <location filename="../qml/symbian/SettingsPage.qml" line="134"/>
@@ -527,15 +529,15 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/sailfish/SettingsPage.qml" line="83"/>
         <source>Automatically start translation after inserting text with &lt;i&gt;Paste&lt;/i&gt; button</source>
-        <translation type="unfinished"></translation>
+        <translation>Αυτόματη εκκίνησης της μετάφρασης κατά την εισαγωγή του κειμένου με το κουμπί &lt;i&gt;Επικόλλησης&lt;/i&gt;</translation>
     </message>
 </context>
 <context>
     <name>TranslationInterface</name>
     <message>
-        <location filename="../src/translationinterface.cpp" line="413"/>
+        <location filename="../src/translationinterface.cpp" line="427"/>
         <source>Please, enter the source text</source>
-        <translation>Εισάγετε το κείμενο πηγής...</translation>
+        <translation>Παρακαλώ, εισαγάγετε το κείμενο πηγής</translation>
     </message>
 </context>
 <context>
@@ -569,7 +571,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../src/translationservice.h" line="134"/>
         <source>%1 service returned an error: &quot;%2&quot;</source>
-        <translation type="unfinished"></translation>
+        <translation>Η υπηρεσία %1 επέστρεψε ένα σφάλμα: «%2»</translation>
     </message>
     <message>
         <location filename="../src/translationservice.h" line="136"/>
@@ -579,7 +581,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../src/translationservice.h" line="138"/>
         <source>Service %1 was unable to translate the text</source>
-        <translation type="unfinished"></translation>
+        <translation>Δεν ήταν δυνατή η μετάφραση του κειμένου από την υπηρεσία %1</translation>
     </message>
 </context>
 <context>
@@ -708,7 +710,11 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
 	&lt;li&gt;&lt;b&gt;Yandex.Translate&lt;/b&gt; - supports translation and language detection.&lt;/li&gt;
 	&lt;li&gt;&lt;b&gt;Yandex.Dictionaries&lt;/b&gt; - supports dictionary with transcription for some languages, synonyms, and reverse translations.&lt;/li&gt;
 &lt;/ul&gt;&lt;p&gt;More services are possible in future.&lt;/p&gt;</source>
-        <translation type="unfinished"></translation>
+        <translation>&lt;p&gt;Το %1 χρησιμοποιεί έγγραμμες υπηρεσίες μετάφρασης για τις μεταφράσεις.&lt;/p&gt;&lt;p&gt;Τρέχουσες υποστηριζόμενες υπηρεσίες:&lt;/p&gt;&lt;ul&gt;	&lt;li&gt;&lt;b&gt;Google Translate&lt;/b&gt; - υποστηρίζει την μετάφραση, την απόδοση των περισσοτέρων μη λατινικών γραφών, τον εντοπισμό της γλώσσας, και λεξικό με αντίστροφες μεταφράσεις ενός λήμματος.&lt;/li&gt;
+	&lt;li&gt;&lt;b&gt;Microsoft Translator&lt;/b&gt; (ή &lt;b&gt;Bing Translator&lt;/b&gt;) - υποστηρίζει μεταφράσεις και εντοπισμό γλώσσας.&lt;/li&gt;
+	&lt;li&gt;&lt;b&gt;Yandex.Translate&lt;/b&gt; - υποστηρίζει μεταφράσεις και εντοπισμό γλώσσας.&lt;/li&gt;
+	&lt;li&gt;&lt;b&gt;Yandex.Dictionaries&lt;/b&gt; - υποστήριξη λεξικού και απομαγνητοφώνηση για μερικές γλώσσες, συνώνυμα, και αντίστροφη μετάφραση.&lt;/li&gt;
+&lt;/ul&gt;&lt;p&gt;Πιθανώς να προστεθούν και άλλες υπηρεσίες μελλοντικά.&lt;/p&gt;</translation>
     </message>
 </context>
 <context>
@@ -716,7 +722,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/bb10/main.qml" line="48"/>
         <source>About</source>
-        <translation>Σχετικα με</translation>
+        <translation>Περί</translation>
     </message>
     <message>
         <location filename="../qml/bb10/main.qml" line="56"/>
@@ -746,7 +752,7 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/bb10/main.qml" line="115"/>
         <source>How much would you like to donate?</source>
-        <translation type="unfinished"></translation>
+        <translation>Τι ποσό επιθυμείτε να δωρίσετε;</translation>
     </message>
     <message>
         <location filename="../qml/bb10/main.qml" line="116"/>
@@ -756,17 +762,17 @@ However, if it&apos;s not enough, the amount of your donation along with &quot;T
     <message>
         <location filename="../qml/bb10/main.qml" line="134"/>
         <source>Payment Succeeded</source>
-        <translation type="unfinished"></translation>
+        <translation>Η πληρωμή ήταν επιτυχής</translation>
     </message>
     <message>
         <location filename="../qml/bb10/main.qml" line="135"/>
         <source>Thanks a lot for your donation!</source>
-        <translation type="unfinished"></translation>
+        <translation>Ευχαριστώ πολύ για την δωρεά!</translation>
     </message>
     <message>
         <location filename="../qml/bb10/main.qml" line="139"/>
         <source>Payment Error</source>
-        <translation type="unfinished"></translation>
+        <translation>Σφάλμα πληρωμής</translation>
     </message>
 </context>
 </TS>