+++ src/translationinterface.cpp
@@ -26,6 +26,10 @@
 #include "translationservicesmodel.h"
 #include "languagelistmodel.h"
 #ifdef Q_OS_BLACKBERRY
+#   include <bb/cascades/Application>
+#   include <bb/cascades/ThemeSupport>
+#   include <bb/cascades/Theme>
+#   include <bb/cascades/ColorTheme>
 #   include <bb/system/InvokeManager>
 #   include "bb10/dictionarymodel.h"
 #else
@@ -118,11 +122,13 @@ TranslationInterface::TranslationInterface(QObject *parent)
 # endif
 
     updatePersistentProperties();
+    m_lastKeepAlive = m_settings->value("Analytics/LastKeepAlive", QDate(1970, 1, 1)).toDate();
 
     m_privacyLevel = PrivacyLevel(m_settings->value("PrivacyLevel", UndefinedPrivacy).toInt());
     if (m_privacyLevel != NoPrivacy)
         m_analytics->setPrivacyEnabled(true);
 
+    bool resetKeepAlive = false;
     if (m_settings->contains("AppInfo/CurrentVersion")) {
         const QString version = m_settings->value("AppInfo/CurrentVersion").toString();
         if (version != QCoreApplication::applicationVersion()) {
@@ -130,6 +136,7 @@ TranslationInterface::TranslationInterface(QObject *parent)
             props.insert("Old Version", version);
             props.insert("New Version", QCoreApplication::applicationVersion());
             m_analytics->trackEvent("Upgrade", props);
+            resetKeepAlive = true;
 
             m_settings->setValue("AppInfo/CurrentVersion", QCoreApplication::applicationVersion());
         }
@@ -137,8 +144,15 @@ TranslationInterface::TranslationInterface(QObject *parent)
         QVariantMap props;
         props.insert("Version", QCoreApplication::applicationVersion());
         m_analytics->trackEvent("Installation", props);
+        resetKeepAlive = true;
 
         m_settings->setValue("AppInfo/CurrentVersion", QCoreApplication::applicationVersion());
+    }
+
+    if (resetKeepAlive) {
+        const QDate today(QDate::currentDate());
+        m_settings->setValue("Analytics/LastKeepAlive", today);
+        m_lastKeepAlive = today;
     }
 
     if (m_privacyLevel == NoPrivacy) {
@@ -421,6 +435,8 @@ void TranslationInterface::translate()
         QVariantMap props;
         fillTranslationProperties(props);
         m_analytics->trackEvent("Translation", props);
+    } else {
+        trackKeepAlive();
     }
 #endif
 
@@ -624,15 +640,39 @@ void TranslationInterface::updatePersistentProperties()
         lc = QLocale(QLocale::English, QLocale::UnitedStates);
     props.insert("OS Name", m_analytics->deviceInfo().os.name);
     props.insert("UI Language", lc.name());
-    props.insert("Inverted Theme", getSettingsValue("InvertedTheme"));
-    props.insert("Translate on Enter", getSettingsValue("TranslateOnEnter"));
-    props.insert("Paste'n'Translate", getSettingsValue("TranslateOnPaste"));
+
+    // On BlackBerry 10, the default theme (dark/light) depends on the phone model.
+    // On other systems, the default theme is the same for all models.
+    const bool invertedTheme
+#ifdef Q_OS_BLACKBERRY
+            = bb::cascades::Application::instance()->themeSupport()->theme()->colorTheme()->style()
+              == bb::cascades::VisualStyle::Dark;
+#else
+            = false;
+#endif
+
+    props.insert("Inverted Theme", getSettingsValue("InvertedTheme", invertedTheme).toBool());
+    props.insert("Translate on Enter", getSettingsValue("TranslateOnEnter", false).toBool());
+    props.insert("Paste'n'Translate", getSettingsValue("TranslateOnPaste", true).toBool());
     m_analytics->setPersistentUserProperties(props);
 }
 
 void TranslationInterface::trackSessionStart()
 {
     m_analytics->trackEvent("Session Start", QVariantMap(), true);
+}
+
+void TranslationInterface::trackKeepAlive()
+{
+    const QDate today(QDate::currentDate());
+    if (m_lastKeepAlive.daysTo(today) != 0) {
+        // Send keep-alive once a day so that Amplitude Analytics treats us as an active user
+        if (m_privacyLevel != NoPrivacy)
+            m_analytics->trackEvent("Keep Alive");
+
+        m_settings->setValue("Analytics/LastKeepAlive", today);
+        m_lastKeepAlive = today;
+    }
 }
 #endif